# api/openapi.yaml
openapi: 3.0.3
info:
  title: WalletHub API
  description: |
    WalletHub - платёжный шлюз для управления кошельками и транзакциями.

    ## Основные возможности
    - Управление пользователями и KYC верификация
    - Создание и управление мультивалютными кошельками (Fiat + Crypto)
    - Проведение финансовых операций (deposit, withdraw, transfer)
    - Полная история транзакций

    ## Аутентификация
    API использует Bearer Token аутентификацию. Передайте токен в заголовке:
    ```
    Authorization: Bearer <your-token>
    ```

    ## Идемпотентность
    Все финансовые операции поддерживают идемпотентность через `idempotency_key`.
    При повторном запросе с тем же ключом вернётся результат первой операции.

    ## Rate Limiting
    - Общий лимит: 100 запросов/минуту
    - Финансовые операции: 30 запросов/минуту

    Заголовки ответа:
    - `X-RateLimit-Limit`: Максимум запросов
    - `X-RateLimit-Remaining`: Оставшееся количество
    - `X-RateLimit-Reset`: Unix timestamp сброса

    ## Формат ответа
    Все ответы имеют единый формат:
    ```json
    {
      "success": true/false,
      "data": { ... },
      "error": { "code": "...", "message": "..." },
      "meta": { "page": 1, "total": 100 },
      "request_id": "uuid",
      "timestamp": "2024-01-01T00:00:00Z"
    }
    ```
  version: 1.0.0
  contact:
    name: WalletHub Support
    email: support@wallethub.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
...
paths:
...
  /api/v1/wallets:
    post:
      tags: [Wallets]
      summary: Create wallet
      description: Create a new wallet for a user with specified currency
      operationId: createWallet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWalletRequest'
      responses:
        '201':
          description: Wallet created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/BusinessRuleError'

    get:
      tags: [Wallets]
      summary: List wallets
      description: Get paginated list of wallets with optional filters
      operationId: listWallets
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - name: user_id
          in: query
          schema:
            type: string
            format: uuid
        - name: currency_code
          in: query
          schema:
            type: string
            minLength: 3
            maxLength: 3
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/WalletStatus'
      responses:
        '200':
          description: List of wallets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletListResponse'
  /api/v1/wallets/{id}:
    get:
      tags: [Wallets]
      summary: Get wallet by ID
      description: Get wallet details by UUID
      operationId: getWallet
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Wallet details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/wallets/{id}/credit:
    post:
      tags: [Wallets]
      summary: Credit wallet (deposit)
      description: Add funds to a wallet
      operationId: creditWallet
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditWalletRequest'
      responses:
        '200':
          description: Wallet credited successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletOperationResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConcurrencyError'
        '422':
          $ref: '#/components/responses/BusinessRuleError'
  /api/v1/wallets/{id}/debit:
    post:
      tags: [Wallets]
      summary: Debit wallet (withdraw)
      description: Remove funds from a wallet
      operationId: debitWallet
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DebitWalletRequest'
      responses:
        '200':
          description: Wallet debited successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletOperationResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          description: Insufficient balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/wallets/{id}/transfer:
    post:
      tags: [Wallets]
      summary: Transfer funds
      description: Transfer funds from source wallet to destination wallet
      operationId: transferFunds
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Source wallet ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferFundsRequest'
      responses:
        '200':
          description: Transfer completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResultResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/BusinessRuleError'

  # ============================================
  # Transactions
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ConflictError:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BusinessRuleError:
      description: Business rule violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConcurrencyError:
      description: Concurrency error (retry needed)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ApiMeta:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
    ErrorResponse:
      type: object
      required: [success, error, request_id, timestamp]
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/ApiError'
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: NOT_FOUND
        message:
          type: string
        details:
          type: object
...
    CreateWalletRequest:
      type: object
      required: [user_id, currency_code]
      properties:
        user_id:
          type: string
          format: uuid
        currency_code:
          type: string
          minLength: 3
          maxLength: 3
          example: USD
    CreditWalletRequest:
      type: object
      required: [amount, idempotency_key, description]
      properties:
        amount:
          type: string
          pattern: '^\d+(\.\d{1,8})?$'
          example: "100.50"
        idempotency_key:
          type: string
          format: uuid
        description:
          type: string
          maxLength: 500
        external_reference:
          type: string
    DebitWalletRequest:
      type: object
      required: [amount, idempotency_key, description]
      properties:
        amount:
          type: string
          pattern: '^\d+(\.\d{1,8})?$'
        idempotency_key:
          type: string
          format: uuid
        description:
          type: string
          maxLength: 500
        external_reference:
          type: string
    TransferFundsRequest:
      type: object
      required: [destination_wallet_id, amount, idempotency_key, description]
      properties:
        destination_wallet_id:
          type: string
          format: uuid
        amount:
          type: string
          pattern: '^\d+(\.\d{1,8})?$'
        idempotency_key:
          type: string
          format: uuid
        description:
          type: string
          maxLength: 500
    Wallet:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        currency_code:
          type: string
        wallet_type:
          type: string
          enum: [FIAT, CRYPTO]
        status:
          $ref: '#/components/schemas/WalletStatus'
        available_balance:
          type: string
        pending_balance:
          type: string
        total_balance:
          type: string
        daily_limit:
          type: string
        monthly_limit:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    WalletStatus:
      type: string
      enum: [ACTIVE, SUSPENDED, LOCKED, CLOSED]
    WalletResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Wallet'
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
    WalletListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            wallets:
              type: array
              items:
                $ref: '#/components/schemas/Wallet'
            total_count:
              type: integer
        meta:
          $ref: '#/components/schemas/ApiMeta'
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
    WalletOperationResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            wallet:
              $ref: '#/components/schemas/Wallet'
            transaction_id:
              type: string
              format: uuid
            message:
              type: string
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
    TransferResultResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            source_wallet:
              $ref: '#/components/schemas/Wallet'
            destination_wallet:
              $ref: '#/components/schemas/Wallet'
            transaction_id:
              type: string
              format: uuid
            amount:
              type: string
            status:
              type: string
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time

    # ============================================
    # Transaction Schemas
    # ============================================
...